Fixed various property duplication bugs and allowed files with unreadable properties to still be loaded
When UProperty.(De)Serialize throws a UESerializationException, it'll be handled as if it were an unsupported property type. The GUI will now display the reason for failing to load a property.

Fixed UTextProperty.Clone and UNameProperty.Clone copying strings by reference.

Fixed the Parent field in the Children of a duplicated property not being updated.

When a file fails to load or save, the exception message will now be displayed.